requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.3
comtypes==1.1.10
distlib==0.3.2
dj-database-url==0.5.0
Django==3.1.7
django-embed-video==1.4.1
django-heroku==0.3.1
django-simple-search==1.0.2
docopt==0.6.2
filelock==3.0.12
gunicorn==20.1.0
idna==3.2
Js2Py==0.71
packaging==21.0
Pillow==8.2.0
pipwin==0.5.1
psycopg2==2.9.3
pyjsparser==2.7.1
pyparsing==2.4.7
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
pyttsx3==2.90
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.6
virtualenv==20.4.7
whitenoise==6.0.0
wikipedia==1.4.0
